refactor: Migrate spotify in a Gradle module.
Issue #61
