--- app/src/main/java/com/guru/composecookbook/ui/demoapps/spotify/SpotifyNavType.kt
@@ -1,5 +1,0 @@
-package com.guru.composecookbook.ui.demoapps.spotify
-
-enum class SpotifyNavType {
-    HOME, SEARCH, LIBRARY
-}
