+++ app/src/main/java/com/guru/composecookbook/ui/demoapps/tiktok/home/HomeScreen.kt
@@ -30,8 +30,8 @@ import androidx.compose.ui.unit.sp
 import com.guru.composecookbook.R
 import com.guru.composecookbook.theme.tiktokRed
 import com.guru.composecookbook.theme.typography
-import com.guru.composecookbook.ui.demoapps.spotify.data.Album
-import com.guru.composecookbook.ui.demoapps.spotify.data.SpotifyDataProvider
+import com.guru.composecookbook.spotify.data.models.Album
+import com.guru.composecookbook.spotify.data.SpotifyDataProvider
 import com.guru.composecookbook.ui.demoapps.tiktok.TikTokPlayer
 import com.guru.composecookbook.ui.demoapps.tiktok.TiktokHomeInteractionEvents
 import com.guru.composecookbook.carousel.Pager
