+++ app/src/main/java/com/guru/composecookbook/ui/home/advancelists/SwipeableLists.kt
@@ -29,8 +29,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.guru.composecookbook.theme.green500
 import com.guru.composecookbook.theme.typography
-import com.guru.composecookbook.ui.demoapps.spotify.data.Album
-import com.guru.composecookbook.ui.demoapps.spotify.data.SpotifyDataProvider
+import com.guru.composecookbook.spotify.data.models.Album
+import com.guru.composecookbook.spotify.data.SpotifyDataProvider
 import kotlin.math.roundToInt
 
 @ExperimentalMaterialApi
