+++ app/src/main/java/com/guru/composecookbook/ui/learnwidgets/AppBars.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.unit.dp
 import com.guru.composecookbook.R
 import com.guru.composecookbook.theme.twitterColor
 import com.guru.composecookbook.theme.typography
-import com.guru.composecookbook.ui.demoapps.spotify.SpotifyNavType
+import com.guru.composecookbook.spotify.ui.home.SpotifyNavType
 import com.guru.composecookbook.ui.utils.SubtitleText
 import com.guru.composecookbook.ui.utils.TitleText
