+++ demos/spotify/src/main/java/com/guru/composecookbook/spotify/ui/home/components/PlayerBottomBar.kt
@@ -1,0 +1,61 @@
+package com.guru.composecookbook.spotify.ui.home.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.FavoriteBorder
+import androidx.compose.material.icons.filled.PlayArrow
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.guru.composecookbook.spotify.R
+import com.guru.composecookbook.theme.graySurface
+import com.guru.composecookbook.theme.typography
+
+@Composable
+fun PlayerBottomBar(modifier: Modifier) {
+    val bottomBarHeight = 57.dp
+    val backgroundColor =
+        if (isSystemInDarkTheme()) graySurface else MaterialTheme.colors.background
+    Row(
+        modifier = modifier
+            .padding(bottom = bottomBarHeight)
+            .fillMaxWidth()
+            .background(color = backgroundColor),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.adele21),
+            modifier = Modifier.size(65.dp),
+            contentDescription = null,
+            contentScale = ContentScale.Crop
+        )
+        Text(
+            text = "Someone Like you by Adele",
+            style = typography.h6.copy(fontSize = 14.sp),
+            modifier = Modifier
+                .padding(8.dp)
+                .weight(1f),
+        )
+        Icon(
+            imageVector = Icons.Default.FavoriteBorder, modifier = Modifier.padding(8.dp),
+            contentDescription = null
+        )
+        Icon(
+            imageVector = Icons.Default.PlayArrow, modifier = Modifier.padding(8.dp),
+            contentDescription = null
+        )
+    }
+}
