+++ demos/spotify/src/main/java/com/guru/composecookbook/spotify/ui/home/components/SpotifyHomeGridItem.kt
@@ -1,0 +1,61 @@
+package com.guru.composecookbook.spotify.ui.home.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Card
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.guru.composecookbook.theme.graySurface
+import com.guru.composecookbook.theme.typography
+import com.guru.composecookbook.spotify.data.models.Album
+import com.guru.composecookbook.spotify.data.SpotifyDataProvider
+import com.guru.composecookbook.spotify.ui.details.SpotifyDetailActivity
+
+@Composable
+fun SpotifyHomeGridItem(album: Album) {
+    val cardColor = if (isSystemInDarkTheme()) graySurface else MaterialTheme.colors.background
+    Card(
+        elevation = 4.dp,
+        backgroundColor = cardColor,
+        modifier = Modifier
+            .clip(RoundedCornerShape(8.dp))
+            .padding(8.dp)
+            .clickable(onClick = {})
+    ) {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            Image(
+                painter = painterResource(id = album.imageId),
+                contentDescription = null,
+                modifier = Modifier.size(55.dp),
+                contentScale = ContentScale.Crop
+            )
+            Text(
+                text = album.song,
+                style = typography.h6.copy(fontSize = 14.sp),
+                modifier = Modifier.padding(horizontal = 8.dp)
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun PreviewSpotifyHomeGridItem() {
+    val album = remember { SpotifyDataProvider.album }
+    SpotifyHomeGridItem(album)
+}
