+++ settings.gradle.kts
@@ -11,3 +11,4 @@ include(":components:charts")
 include(":components:tags")
 include(":components:carousel")
 include(":components:verticalgrid")
+include(":demos:spotify")
